Update to pytorch 0.4

+++ tractseg/libs/PytorchUtils.py
@@ -65,7 +65,8 @@ class PytorchUtils:
             intersect = torch.sum(y_true[:, i] * y_pred[:, i])  # works because all multiplied by 0 gets 0
             denominator = torch.sum(y_true[:, i]) + torch.sum(y_pred[:, i])  # works because all multiplied by 0 gets 0
             #maybe have to cast to float here (for python3 ??) otherwise always 0
-            f1 = (2 * intersect) / (denominator + 1e-6)
+            # f1 = (2 * intersect) / (denominator + 1e-6)
+            f1 = (2 * intersect.float()) / (denominator.float() + 1e-6)
             f1s.append(f1)
         if per_class:
             return np.array(f1s)
